requirements.txt
# Core dependencies
ultralytics>=8.2.34
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Data handling
PyYAML>=6.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0